Refactors to "command:again"
